+++ naga/src/front/glsl/context.rs
@@ -409,7 +409,7 @@ impl<'a> Context<'a> {
     /// - If more than one [`StmtContext`] are active at the same time or if the
     ///   previous call didn't use it in lowering.
     #[must_use]
-    pub fn stmt_ctx(&mut self) -> StmtContext {
+    pub const fn stmt_ctx(&mut self) -> StmtContext {
         self.stmt_ctx.take().unwrap()
     }