+++ wgpu-hal/src/dx12/instance.rs
@@ -37,7 +37,6 @@ impl crate::Instance for super::Instance {
                     .flags
                     .intersects(wgt::InstanceFlags::GPU_BASED_VALIDATION)
                 {
-                    #[allow(clippy::collapsible_if)]
                     if let Ok(debug1) = debug_controller.cast::<Direct3D12::ID3D12Debug1>() {
                         unsafe { debug1.SetEnableGPUBasedValidation(true) }
                     } else {