+++ wgpu-hal/src/vulkan/command.rs
@@ -284,7 +284,7 @@ impl crate::CommandEncoder for super::CommandEncoder {
         if self.device.workarounds.contains(
             super::Workarounds::FORCE_FILL_BUFFER_WITH_SIZE_GREATER_4096_ALIGNED_OFFSET_16,
         ) && range_size >= 4096
-            && range.start % 16 != 0
+            && !range.start.is_multiple_of(16)
         {
             let rounded_start = wgt::math::align_to(range.start, 16);
             let prefix_size = rounded_start - range.start;