+++ CHANGELOG.md
@@ -48,6 +48,14 @@ Bottom level categories:
 - Added support for no-perspective barycentric coordinates. By @atlv24 in [#8852](https://github.com/gfx-rs/wgpu/issues/8852).
 - Added support for obtaining `AdapterInfo` from `Device`. By @sagudev in [#8807](https://github.com/gfx-rs/wgpu/pull/8807).
 - Added `Limits::or_worse_values_from`. By @atlv24 in [#8870](https://github.com/gfx-rs/wgpu/pull/8870).
+- Made the following available in `const` contexts:
+    - `naga`
+        - `Arena::len`
+        - `Arena::is_empty`
+        - `Range::first_and_last`
+        - `front::wgsl::Frontend::set_options`
+        - `ir::Block::is_empty`
+        - `ir::Block::len`
 
 #### Naga